Support for Math Expressions #9
. Implemented logic for executing math
expressions
. In short, mathematical expressions are
modelled as a binary tree, where some nodes
have higher precedence than others. When building
the tree - out of a math expression - we reorganize
it as nodes are added - where a node can be a number,
an operator, or another tree (for expressions grouped
between '(' ')'. Nodes with higher precedence are never
touched though, to keep operator and parenthesis
precedence
. Added instructions for each operation (+, -, *, and /)
. Added a .knit file to test that