Refactoring on the compiler
. Refactored how variable refences and constants
are modelled in the language definition. Also
changed how the compiler handles them. Now there
are specific contexts for those, responsible for
pushing onto the operand stack either a constant or
the value of a variable.
. Deleted all the *Value classes, their function is now
handled by a constant context
. Made all contexts that use constants and variable refences
a composite context, where the corresponding context
will be added to during compilation